requirements.txt
langchain
langchain-community
langchain-openai
langchan-groq
langchain-experimental
streamlit
pandas
matplotlib
seaborn
pygwalker
tiktoken
tabulate
python-dotenv